Интерфейс Input [#181092 #330613]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -5,5 +5,5 @@ public interface Transport {
 
     void passengers(int passenger);
 
-    double refuel(int fuel);
+    int refuel(int fuel);
 }
